+++ cypress/integration/racing_car_input.spec.js
@@ -1,6 +1,6 @@
 /// <reference types="Cypress" />
 
-import { ERROR_MESSAGE } from '../../src/js/util/errorMessage.js';
+import { ERROR_MESSAGE } from '../../src/js/util/message.js';
 
 describe('Racing Car 게임 : INPUT 검증 테스트', () => {
   beforeEach(() => {
