+++ src/js/util/message.js
